move data generator from utils to idata w dates
