+++ forecasttools/idata_w_dates_to_df.py
@@ -12,8 +12,64 @@ from datetime import datetime, timedelta
 import arviz as az
 import numpy as np
 import polars as pl
+import xarray as xr
 
 import forecasttools
+
+
+def generate_time_range_for_dim(
+    start_time_as_dt: datetime,
+    variable_data: xr.DataArray,
+    dimension: str,
+    time_step: timedelta,
+):
+    """
+    Generates a range of times based on the
+    start date, time step, and variable's
+    dimension size. A range of dates is
+    generated if the start date is a date and
+    the time step is in days. A range of times
+    is generated if the start date is a time
+    and/or the time step is a time.
+    """
+
+    # get the size of the dimension
+    interval_size = variable_data.sizes[dimension]
+    # number of seconds in a day
+    SECONDS_IN_DAY = timedelta(days=1).total_seconds()
+
+    # total number of seconds in the time_step
+    total_seconds = time_step.total_seconds()
+
+    # determine the interval string for Polars
+    if (
+        total_seconds % SECONDS_IN_DAY == 0
+    ):  # check if time_step is in full days
+        # use date_range for dates
+        return (
+            pl.date_range(
+                start=start_time_as_dt,
+                end=start_time_as_dt + (interval_size - 1) * time_step,
+                interval=time_step,  # use the calculated interval
+                closed="both",
+                eager=True,  # return a Polars Series
+            )
+            .to_numpy()
+            .astype("datetime64[D]")  # date format
+        )
+    else:
+        # use datetime_range for times
+        return (
+            pl.datetime_range(
+                start=start_time_as_dt,
+                end=start_time_as_dt + (interval_size - 1) * time_step,
+                interval=time_step,  # use the calculated interval
+                closed="both",
+                eager=True,  # return a Polars Series
+            )
+            .to_numpy()
+            .astype("datetime64[ns]")  # time format
+        )
 
 
 def add_time_coords_to_idata_dimension(
@@ -90,7 +146,7 @@ def add_time_coords_to_idata_dimension(
     forecasttools.validate_idata_group_var_dim(
         variable_data=variable_data, dimension=dimension
     )
-    interval_dates = forecasttools.generate_time_range_for_dim(
+    interval_dates = generate_time_range_for_dim(
         start_time_as_dt=start_time_as_dt,
         variable_data=variable_data,
         dimension=dimension,
